Organized AsyncExecutionApiAdapter implementations into files
