+++ src/NUnitFramework/tests/AsyncExecutionApiAdapter.cs
@@ -22,13 +22,7 @@
 // ***********************************************************************
 
 #if ASYNC
-using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
-using NUnit.Framework.Constraints;
-using NUnit.Framework.Internal;
-using NUnit.Framework.Internal.Builders;
-using NUnit.TestUtilities;
 
 namespace NUnit.Framework
 {
@@ -36,7 +30,7 @@ namespace NUnit.Framework
     /// Adapts each NUnit API that executes async user code synchronously to a common interface
     /// to be tested for behavior.
     /// </summary>
-    public abstract class AsyncExecutionApiAdapter
+    public abstract partial class AsyncExecutionApiAdapter
     {
         public static IEnumerable<AsyncExecutionApiAdapter> All { get; } = new AsyncExecutionApiAdapter[]
         {
@@ -56,239 +50,6 @@ namespace NUnit.Framework
 
         public abstract void Execute(AsyncTestDelegate asyncUserCode);
         public abstract override string ToString();
-
-        private sealed class TaskReturningTestMethodAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                TestBuilder.RunTest(
-                    new NUnitTestFixtureBuilder().BuildFrom(typeof(Fixture), new TestFixtureData(asyncUserCode))
-                ).AssertPassed();
-            }
-
-            private sealed class Fixture
-            {
-                private readonly AsyncTestDelegate _asyncUserCode;
-
-                public Fixture(AsyncTestDelegate asyncUserCode)
-                {
-                    _asyncUserCode = asyncUserCode;
-                }
-
-                [Test]
-                public Task TestMethod() => _asyncUserCode.Invoke();
-            }
-
-            public override string ToString() => "[Test] Task TestMethod() { … }";
-        }
-
-        private sealed class TaskReturningSetUpAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                TestBuilder.RunTest(
-                    new NUnitTestFixtureBuilder().BuildFrom(typeof(Fixture), new TestFixtureData(asyncUserCode))
-                ).AssertPassed();
-            }
-
-            private sealed class Fixture
-            {
-                private readonly AsyncTestDelegate _asyncUserCode;
-
-                public Fixture(AsyncTestDelegate asyncUserCode)
-                {
-                    _asyncUserCode = asyncUserCode;
-                }
-                
-                [SetUp]
-                public Task SetUp() => _asyncUserCode.Invoke();
-                
-                [Test]
-                public void DummyTest() { }
-            }
-
-            public override string ToString() => "[SetUp] Task SetUp() { … }";
-        }
-
-        private sealed class TaskReturningTearDownAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                TestBuilder.RunTest(
-                    new NUnitTestFixtureBuilder().BuildFrom(typeof(Fixture), new TestFixtureData(asyncUserCode))
-                ).AssertPassed();
-            }
-
-            private sealed class Fixture
-            {
-                private readonly AsyncTestDelegate _asyncUserCode;
-
-                public Fixture(AsyncTestDelegate asyncUserCode)
-                {
-                    _asyncUserCode = asyncUserCode;
-                }
-
-                [Test]
-                public void DummyTest() { }
-
-                [TearDown]
-                public Task TearDown() => _asyncUserCode.Invoke();
-            }
-
-            public override string ToString() => "[TearDown] Task TearDown() { … }";
-        }
-
-        private sealed class TaskReturningOneTimeSetUpAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                TestBuilder.RunTest(
-                    new NUnitTestFixtureBuilder().BuildFrom(typeof(Fixture), new TestFixtureData(asyncUserCode))
-                ).AssertPassed();
-            }
-
-            private sealed class Fixture
-            {
-                private readonly AsyncTestDelegate _asyncUserCode;
-
-                public Fixture(AsyncTestDelegate asyncUserCode)
-                {
-                    _asyncUserCode = asyncUserCode;
-                }
-
-                [OneTimeSetUp]
-                public Task OneTimeSetUp() => _asyncUserCode.Invoke();
-
-                [Test]
-                public void DummyTest() { }
-            }
-
-            public override string ToString() => "[OneTimeSetUp] Task OneTimeSetUp() { … }";
-        }
-
-        private sealed class TaskReturningOneTimeTearDownAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                TestBuilder.RunTest(
-                    new NUnitTestFixtureBuilder().BuildFrom(typeof(Fixture), new TestFixtureData(asyncUserCode))
-                ).AssertPassed();
-            }
-
-            private sealed class Fixture
-            {
-                private readonly AsyncTestDelegate _asyncUserCode;
-
-                public Fixture(AsyncTestDelegate asyncUserCode)
-                {
-                    _asyncUserCode = asyncUserCode;
-                }
-
-                [Test]
-                public void DummyTest() { }
-
-                [OneTimeTearDown]
-                public Task OneTimeTearDown() => _asyncUserCode.Invoke();
-            }
-
-            public override string ToString() => "[OneTimeTearDown] Task OneTimeTearDown() { … }";
-        }
-
-        private sealed class AssertThrowsAsyncAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                var ex = (Exception)null;
-                using (new TestExecutionContext.IsolatedContext())
-                {
-                    try
-                    {
-                        ex = Assert.ThrowsAsync<Exception>(asyncUserCode);
-                    }
-                    catch { }
-                }
-
-                if (ex != null) ExceptionHelper.Rethrow(ex);
-            }
-
-            public override string ToString() => "Assert.ThrowsAsync(…)";
-        }
-
-        private sealed class AssertDoesNotThrowAsyncAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                Assert.DoesNotThrowAsync(asyncUserCode);
-            }
-
-            public override string ToString() => "Assert.DoesNotThrowAsync(…)";
-        }
-
-        private sealed class AssertCatchAsyncAdapter : AsyncExecutionApiAdapter
-        {
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                var ex = (Exception)null;
-                using (new TestExecutionContext.IsolatedContext())
-                {
-                    try
-                    {
-                        ex = Assert.CatchAsync(asyncUserCode);
-                    }
-                    catch { }
-                }
-
-                if (ex != null) ExceptionHelper.Rethrow(ex);
-            }
-
-            public override string ToString() => "Assert.CatchAsync(…)";
-        }
-
-        private abstract class ConstraintApiAdapter : AsyncExecutionApiAdapter
-        {
-            private readonly Constraint _constraint;
-
-            protected ConstraintApiAdapter(Constraint constraint)
-            {
-                _constraint = constraint;
-            }
-
-            public override void Execute(AsyncTestDelegate asyncUserCode)
-            {
-                _constraint.ApplyTo(asyncUserCode);
-                _constraint.ApplyTo(asyncUserCode.Invoke); // ActualValueDelegate<> overload
-            }
-
-            public sealed override string ToString() => _constraint.GetType().Name + ".ApplyTo(…)";
-        }
-
-        private sealed class ThrowsNothingConstraintAdapter : ConstraintApiAdapter
-        {
-            public ThrowsNothingConstraintAdapter() : base(new ThrowsNothingConstraint())
-            {
-            }
-        }
-
-        private sealed class ThrowsConstraintAdapter : ConstraintApiAdapter
-        {
-            public ThrowsConstraintAdapter() : base(new ThrowsConstraint(DummyConstraint.Instance))
-            {
-            }
-        }
-
-        private sealed class ThrowsExceptionConstraintAdapter : ConstraintApiAdapter
-        {
-            public ThrowsExceptionConstraintAdapter() : base(new ThrowsExceptionConstraint())
-            {
-            }
-        }
-
-        private sealed class NormalConstraintAdapter : ConstraintApiAdapter
-        {
-            public NormalConstraintAdapter() : base(DummyConstraint.Instance)
-            {
-            }
-        }
     }
 }
 #endif
